requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.12
dj-rest-auth==2.1.11
Django==3.2.6
django-allauth==0.45.0
django-cors-headers==3.8.0
django-filter==2.4.0
django-oauth-toolkit==1.5.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
future==0.18.2
futures==3.1.1
goslate==1.5.1
idna==3.2
joblib==1.0.1
jwcrypto==1.0
Markdown==3.3.4
nltk==3.6.2
oauthlib==3.1.1
pycparser==2.20
PyDictionary==2.0.1
PyJWT==2.1.0
python-whois==0.7.3
python3-openid==3.2.0
pytz==2021.1
regex==2021.8.21
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tqdm==4.62.1
urllib3==1.26.6
wrapt==1.12.1
